Add files via upload
Note: code that is commented out (#...), specifically, by "#***" at the end of a line (for single comments) or above and below (for distinguishing commented-out blocks of code). Until I update this code when complete, you must uncomment the above for the first run. If there is something commented out that is not  demarcated by "#***", then it's meant to be commented out. 

This code is incomplete. I am putting this code up so others can see what kind of progress I've made and where I'm going. I do a decent job tracking what is done and where I intend to go in my scripts, so others can get a gist of my competency in R.